requirements.txt
asgiref==3.5.2
bokeh==2.4.3
Django==4.1.3
django-csp==3.7
django-extensions==3.2.1
django-filter==22.1
django-htmx==1.13.0
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
packaging==21.3
Pillow==9.3.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
PyYAML==6.0
sqlparse==0.4.3
tornado==6.2
typing_extensions==4.4.0
setuptools==65.5.1


# testing
coverage==6.3.2
pip-audit==2.1.1


#deploy
click==8.1.3
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
pytz==2022.6
tzdata==2022.6
uvicorn==0.19.0
uvloop==0.17.0
websockets==10.4
